+++ Expanded Cheat Menu/Code/Menus/MissionFunc.lua
@@ -367,104 +367,104 @@ function ChoGGi.MenuFuncs.SetSponsor()
 	}
 end
 
-function ChoGGi.MenuFuncs.SetSponsorBonus()
-	local UserSettings = ChoGGi.UserSettings
-	local Presets = Presets
+--~ function ChoGGi.MenuFuncs.SetSponsorBonus()
+--~ 	local UserSettings = ChoGGi.UserSettings
+--~ 	local Presets = Presets
 
-	local item_list = {}
-	local c = 0
-	local objs = Presets.MissionSponsorPreset.Default or ""
-	for i = 1, #objs do
-		local spon = objs[i]
-		if spon.id ~= "random" and spon.id ~= "None" then
-			local descr = GetSponsorDescr(spon, false, "include rockets", true, true)
-			local stats
-			-- the one we want is near the end, but there's also a blank item below it
-			for j = 1, #descr do
-				local des = descr[j]
-				if type(des) == "table" then
-					stats = des
-				end
-			end
+--~ 	local item_list = {}
+--~ 	local c = 0
+--~ 	local objs = Presets.MissionSponsorPreset.Default or ""
+--~ 	for i = 1, #objs do
+--~ 		local spon = objs[i]
+--~ 		if spon.id ~= "random" and spon.id ~= "None" then
+--~ 			local descr = GetSponsorDescr(spon, false, "include rockets", true, true)
+--~ 			local stats
+--~ 			-- the one we want is near the end, but there's also a blank item below it
+--~ 			for j = 1, #descr do
+--~ 				local des = descr[j]
+--~ 				if type(des) == "table" then
+--~ 					stats = des
+--~ 				end
+--~ 			end
 
-			local user_set = UserSettings["Sponsor" .. spon.id]
-			if user_set then
-				user_set = ": " .. tostring(user_set)
-			else
-				user_set = " false"
-			end
-			local save_in = ""
-			if spon.save_in and spon.save_in ~= "" then
-				save_in = "\nsave_in: " .. spon.save_in
-			end
+--~ 			local user_set = UserSettings["Sponsor" .. spon.id]
+--~ 			if user_set then
+--~ 				user_set = ": " .. tostring(user_set)
+--~ 			else
+--~ 				user_set = " false"
+--~ 			end
+--~ 			local save_in = ""
+--~ 			if spon.save_in and spon.save_in ~= "" then
+--~ 				save_in = "\nsave_in: " .. spon.save_in
+--~ 			end
 
-			c = c + 1
-			item_list[c] = {
-				text = Translate(spon.display_name),
-				value = spon.id,
-				hint = Translate(T{spon.effect, stats[2]}) .. "\n\n"
-					.. Strings[302535920001165--[[Enabled Status]]] .. user_set .. save_in,
-			}
-		end
-	end
+--~ 			c = c + 1
+--~ 			item_list[c] = {
+--~ 				text = Translate(spon.display_name),
+--~ 				value = spon.id,
+--~ 				hint = Translate(T{spon.effect, stats[2]}) .. "\n\n"
+--~ 					.. Strings[302535920001165--[[Enabled Status]]] .. user_set .. save_in,
+--~ 			}
+--~ 		end
+--~ 	end
 
-	local function CallBackFunc(choice)
-		if choice.nothing_selected then
-			return
-		end
-		if choice[1].check2 then
-			for i = 1, #item_list do
-				local value = item_list[i].value
-				if type(value) == "string" then
-					value = "Sponsor" .. value
-					UserSettings[value] = nil
-				end
-			end
-		else
-			for i = 1, #choice do
-				local value = choice[i].value
-				for j = 1, #item_list do
-					-- check to make sure it isn't a fake name (no sense in saving it)
-					if item_list[j].value == value and type(value) == "string" then
-						local name = "Sponsor" .. value
-						if choice[1].check1 then
-							UserSettings[name] = nil
-						else
-							UserSettings[name] = true
-						end
-						if UserSettings[name] then
-							ChoGGi.ComFuncs.SetSponsorBonuses(value)
-						end
-					end
-				end
-			end
-		end
+--~ 	local function CallBackFunc(choice)
+--~ 		if choice.nothing_selected then
+--~ 			return
+--~ 		end
+--~ 		if choice[1].check2 then
+--~ 			for i = 1, #item_list do
+--~ 				local value = item_list[i].value
+--~ 				if type(value) == "string" then
+--~ 					value = "Sponsor" .. value
+--~ 					UserSettings[value] = nil
+--~ 				end
+--~ 			end
+--~ 		else
+--~ 			for i = 1, #choice do
+--~ 				local value = choice[i].value
+--~ 				for j = 1, #item_list do
+--~ 					-- check to make sure it isn't a fake name (no sense in saving it)
+--~ 					if item_list[j].value == value and type(value) == "string" then
+--~ 						local name = "Sponsor" .. value
+--~ 						if choice[1].check1 then
+--~ 							UserSettings[name] = nil
+--~ 						else
+--~ 							UserSettings[name] = true
+--~ 						end
+--~ 						if UserSettings[name] then
+--~ 							ChoGGi.ComFuncs.SetSponsorBonuses(value)
+--~ 						end
+--~ 					end
+--~ 				end
+--~ 			end
+--~ 		end
 
-		ChoGGi.SettingFuncs.WriteSettings()
-		MsgPopup(
-			ChoGGi.ComFuncs.SettingState(#choice),
-			Strings[302535920000714--[[Set Bonuses Sponsor]]]
-		)
-	end
+--~ 		ChoGGi.SettingFuncs.WriteSettings()
+--~ 		MsgPopup(
+--~ 			ChoGGi.ComFuncs.SettingState(#choice),
+--~ 			Strings[302535920000714--[[Set Bonuses Sponsor]]]
+--~ 		)
+--~ 	end
 
-	ChoGGi.ComFuncs.OpenInListChoice{
-		callback = CallBackFunc,
-		items = item_list,
-		title = Strings[302535920000714--[[Set Bonuses Sponsor]]],
-		hint = Strings[302535920000106--[[Current]]] .. ": " .. Translate(GetMissionSponsor().display_name) .. "\n\n" .. Strings[302535920001168--[[Modded ones are mostly ignored for now (just cargo space/research points).]]],
-		multisel = true,
-		checkboxes = {
-			{
-				title = Strings[302535920001169--[[Turn Off]]],
-				hint = Strings[302535920001170--[[Turn off selected bonuses (defaults to turning on).]]],
-			},
-			{
-				title = Strings[302535920001171--[[Turn All Off]]],
-				hint = Strings[302535920001172--[[Turns off all bonuses.]]],
-			},
-		},
-	}
-end
+--~ 	ChoGGi.ComFuncs.OpenInListChoice{
+--~ 		callback = CallBackFunc,
+--~ 		items = item_list,
+--~ 		title = Strings[302535920000714--[[Set Bonuses Sponsor]]],
+--~ 		hint = Strings[302535920000106--[[Current]]] .. ": " .. Translate(GetMissionSponsor().display_name) .. "\n\n" .. Strings[302535920001168--[[Modded ones are mostly ignored for now (just cargo space/research points).]]],
+--~ 		multisel = true,
+--~ 		checkboxes = {
+--~ 			{
+--~ 				title = Strings[302535920001169--[[Turn Off]]],
+--~ 				hint = Strings[302535920001170--[[Turn off selected bonuses (defaults to turning on).]]],
+--~ 			},
+--~ 			{
+--~ 				title = Strings[302535920001171--[[Turn All Off]]],
+--~ 				hint = Strings[302535920001172--[[Turns off all bonuses.]]],
+--~ 			},
+--~ 		},
+--~ 	}
+--~ end
 
 function ChoGGi.MenuFuncs.SetCommander()
 	local g_CurrentMissionParams = g_CurrentMissionParams
@@ -524,87 +524,87 @@ function ChoGGi.MenuFuncs.SetCommander()
 	}
 end
 
-function ChoGGi.MenuFuncs.SetCommanderBonus()
-	local Presets = Presets
-	local UserSettings = ChoGGi.UserSettings
+--~ function ChoGGi.MenuFuncs.SetCommanderBonus()
+--~ 	local Presets = Presets
+--~ 	local UserSettings = ChoGGi.UserSettings
 
-	local item_list = {}
-	local c = 0
-	local objs = Presets.CommanderProfilePreset.Default or ""
-	for i = 1, #objs do
-		local comm = objs[i]
-		if comm.id ~= "random" and comm.id ~= "None" then
-			local user_set = UserSettings["Commander" .. comm.id]
+--~ 	local item_list = {}
+--~ 	local c = 0
+--~ 	local objs = Presets.CommanderProfilePreset.Default or ""
+--~ 	for i = 1, #objs do
+--~ 		local comm = objs[i]
+--~ 		if comm.id ~= "random" and comm.id ~= "None" then
+--~ 			local user_set = UserSettings["Commander" .. comm.id]
 
-			c = c + 1
-			item_list[c] = {
-				text = Translate(comm.display_name),
-				value = comm.id,
-				hint = Translate(comm.effect) .. "\n\n"
-					.. Strings[302535920001165--[[Enabled Status]]]
-					.. (user_set and ": " .. user_set or " false"),
-			}
-		end
-	end
+--~ 			c = c + 1
+--~ 			item_list[c] = {
+--~ 				text = Translate(comm.display_name),
+--~ 				value = comm.id,
+--~ 				hint = Translate(comm.effect) .. "\n\n"
+--~ 					.. Strings[302535920001165--[[Enabled Status]]]
+--~ 					.. (user_set and ": " .. comm.id or " false"),
+--~ 			}
+--~ 		end
+--~ 	end
 
-	local function CallBackFunc(choice)
-		if choice.nothing_selected then
-			return
-		end
+--~ 	local function CallBackFunc(choice)
+--~ 		if choice.nothing_selected then
+--~ 			return
+--~ 		end
 
-		if choice[1].check2 then
-			for i = 1, #item_list do
-				local value = item_list[i].value
-				if type(value) == "string" then
-					value = "Commander" .. value
-					UserSettings[value] = nil
-				end
-			end
-		else
-			for i = 1, #choice do
-				for j = 1, #item_list do
-					-- check to make sure it isn't a fake name (no sense in saving it)
-					local value = choice[i].value
-					if item_list[j].value == value and type(value) == "string" then
-						local name = "Commander" .. value
-						if choice[1].check1 then
-							UserSettings[name] = nil
-						else
-							UserSettings[name] = true
-						end
-						if UserSettings[name] then
-							ChoGGi.ComFuncs.SetCommanderBonuses(value)
-						end
-					end
-				end
-			end
-		end
+--~ 		if choice[1].check2 then
+--~ 			for i = 1, #item_list do
+--~ 				local value = item_list[i].value
+--~ 				if type(value) == "string" then
+--~ 					value = "Commander" .. value
+--~ 					UserSettings[value] = nil
+--~ 				end
+--~ 			end
+--~ 		else
+--~ 			for i = 1, #choice do
+--~ 				for j = 1, #item_list do
+--~ 					-- check to make sure it isn't a fake name (no sense in saving it)
+--~ 					local value = choice[i].value
+--~ 					if item_list[j].value == value and type(value) == "string" then
+--~ 						local name = "Commander" .. value
+--~ 						if choice[1].check1 then
+--~ 							UserSettings[name] = nil
+--~ 						else
+--~ 							UserSettings[name] = true
+--~ 						end
+--~ 						if UserSettings[name] then
+--~ 							ChoGGi.ComFuncs.SetCommanderBonuses(value)
+--~ 						end
+--~ 					end
+--~ 				end
+--~ 			end
+--~ 		end
 
-		ChoGGi.SettingFuncs.WriteSettings()
-		MsgPopup(
-			ChoGGi.ComFuncs.SettingState(#choice),
-			Strings[302535920000718--[[Set Bonuses Commander]]]
-		)
-	end
+--~ 		ChoGGi.SettingFuncs.WriteSettings()
+--~ 		MsgPopup(
+--~ 			ChoGGi.ComFuncs.SettingState(#choice),
+--~ 			Strings[302535920000718--[[Set Bonuses Commander]]]
+--~ 		)
+--~ 	end
 
-	ChoGGi.ComFuncs.OpenInListChoice{
-		callback = CallBackFunc,
-		items = item_list,
-		title = Strings[302535920000718--[[Set Bonuses Commander]]],
-		hint = Strings[302535920000106--[[Current]]] .. ": " .. Translate(GetCommanderProfile().display_name),
-		multisel = true,
-		checkboxes = {
-			{
-				title = Strings[302535920001169--[[Turn Off]]],
-				hint = Strings[302535920001170--[[Turn off selected bonuses (defaults to turning on).]]],
-			},
-			{
-				title = Strings[302535920001171--[[Turn All Off]]],
-				hint = Strings[302535920001172--[[Turns off all bonuses.]]],
-			},
-		},
-	}
-end
+--~ 	ChoGGi.ComFuncs.OpenInListChoice{
+--~ 		callback = CallBackFunc,
+--~ 		items = item_list,
+--~ 		title = Strings[302535920000718--[[Set Bonuses Commander]]],
+--~ 		hint = Strings[302535920000106--[[Current]]] .. ": " .. Translate(GetCommanderProfile().display_name),
+--~ 		multisel = true,
+--~ 		checkboxes = {
+--~ 			{
+--~ 				title = Strings[302535920001169--[[Turn Off]]],
+--~ 				hint = Strings[302535920001170--[[Turn off selected bonuses (defaults to turning on).]]],
+--~ 			},
+--~ 			{
+--~ 				title = Strings[302535920001171--[[Turn All Off]]],
+--~ 				hint = Strings[302535920001172--[[Turns off all bonuses.]]],
+--~ 			},
+--~ 		},
+--~ 	}
+--~ end
 
 function ChoGGi.MenuFuncs.ChangeGameLogo()
 	local MissionLogoPresetMap = MissionLogoPresetMap