+++ Mods ChoGGi/Constant Disasters/metadata.lua
@@ -15,7 +15,8 @@ return PlaceObj("ModDef", {
 	"has_options", true,
 	"TagGameplay", true,
 	"TagOther", true,
-	"description", [[Set a minimum interval between disasters (if you want hourly toxic rains for some reason).
+	"description", [[
+Set a minimum interval between disasters (if you want hourly toxic rains for some reason).
 
 This is "separate" from the regular disaster happenings, so turning off all the options won't disable regularly scheduled disasters.
 It won't screw you over and double up the same disaster.
@@ -24,5 +25,6 @@ Mod Options:
 Set all disasters, and the time between (max is 600 hours = 25 Sols, more/less?).
 
 
-Kinda requested by Paladinboyd.]],
+Kinda requested by Paladinboyd.
+]],
 })