+++ Mods ChoGGi/Enable Cheats Pane/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2618713139",
 	"pops_any_uuid", "889e2363-a2e4-4379-be3c-e33d2d87e5e1",
 	"lua_revision", 1007000, -- Picard
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {