+++ Mods ChoGGi/Meteoroid Deposits/items.lua
@@ -1,6 +1,6 @@
 -- See LICENSE for terms
 
-return {
+local props = {
 	PlaceObj("ModItemOptionToggle", {
 		"name", "EnableMod",
 		"DisplayName", T(302535920011303, "<color ChoGGi_yellow>Enable Mod</color>"),
@@ -51,19 +51,23 @@ return {
 		"MinValue", 1,
 		"MaxValue", 50,
 	}),
-	PlaceObj("ModItemOptionToggle", {
+}
+
+if g_AvailableDlc.picard then
+	props[#props+1] = PlaceObj("ModItemOptionToggle", {
 		"name", "ExoticMinerals",
 		"DisplayName", T(608869515243, "Exotic Minerals"),
 		"Help", T(302535920012090, "Drop mineral deposits."),
 		"DefaultValue", false,
-	}),
-	PlaceObj("ModItemOptionNumber", {
+	})
+	props[#props+1] = PlaceObj("ModItemOptionNumber", {
 		"name", "ExoticMineralsThreshold",
 		"DisplayName", T(302535920012091, "Exotic Minerals Threshold"),
 		"Help", T(302535920012092, "Threshold for a new Exotic Minerals deposit."),
 		"DefaultValue", 1,
 		"MinValue", 1,
 		"MaxValue", 50,
-	}),
+	})
+end
 
-}
+return props